Ajustes para remoção de importes inuteis; Ajustes de rotas; Ajustes de pastas e alterada estrutura para usar o basico de signals
