+++ src/app/app.routes.ts
@@ -1,14 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { ContatosComponent } from './contatos/contatos.component'; // Componente para a lista de contatos
+import { ContatosComponent } from './pages/contatos/contatos.component'; // Componente para a lista de contatos
 
 export const routes: Routes = [
+  {path: '', pathMatch: 'full', redirectTo: 'contatos'},
   { path: 'contatos', component: ContatosComponent }, // Rota para a lista de contatos
   { path: '**', redirectTo: '' } // Redireciona para a página inicial em caso de rota não encontrada
 ];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)], // Importa as rotas
-  exports: [RouterModule] // Exporta o RouterModule para que possa ser utilizado em outros módulos
-})
-export class AppRoutingModule {}
