+++ src/app/interfaces/contato.interface.ts
@@ -1,0 +1,6 @@
+export interface Contato {
+    id: number;
+    nome: string;
+    email: string;
+    telefone: string;
+  }
