+++ src/app/pages/contatos/contatos.component.html
@@ -31,7 +31,7 @@
 
   <mat-list>
     <h3 mat-subheader>Contatos</h3>
-    <mat-list-item *ngFor="let contato of contatos">
+    <mat-list-item *ngFor="let contato of contatos()">
       {{ contato.nome }} - {{ contato.email }} - {{ contato.telefone }}
     </mat-list-item>
   </mat-list>
